+++ test/functional-tests/node-transformers/simplifying-transformers/if-statement-simplify-transformer/fixtures/const-variable-declaration-as-prohibited-single-statement.js
@@ -1,0 +1,3 @@
+if (true) {
+    const foo = 1;
+}