+++ test/functional-tests/node-transformers/simplifying-transformers/if-statement-simplify-transformer/fixtures/partial-consequent-only-no-return-single-statement.js
@@ -1,5 +1,5 @@
 function foo () {
     if (true) {
-        const bar = baz();
+        var bar = baz();
     }
 }